Added endpoints DI extensions
The endpoints extensions have been added to the modules. This also includes the removal of some redundant code from Core.Infrastructure.Extensions.cs and Bootstrapper/Program.cs, as now those functionalities are provided through AddEndpoints and UseEndpoints extensions methods. A new EndpointsExtensions.cs file was also created as a part of this change.