Added User configuration + new migration
This commit adds a User entity to the Management module along with necessary infrastructure. A new migration file is introduced to create the User table in the database. Also, the new ValueConverter for encoding and decoding the Password field is implemented. Corresponding changes were made in the `UserConfiguration` and `ManagementDbContextModelSnapshot` files.