Refactored Core.Application structure + added communication related abstractions
The directory structure is refactored to move 'Exceptions', 'Auth', and 'Database' directories out of 'Abstractions'. References in different files are updated accordingly. Additionally, new communication related command and query interfaces and their handler interfaces are introduced in the Core.Application layer to facilitate the handling of different commands and queries in the application.