WIP: Added email service and email related functionalities
This commit adds an email service in Services.Email along with a queue mechanism and sender logic. It also includes Scriban-based email templates for activation of institution accounts, in both English and Polish versions. Moreover, a reference to the MailKit package has been added to the Core.Application project to support the email functionality.