Added GetCurrentUserQuery feature + created interface for ManagementDbContext
This update adds a new GetCurrent user feature that enables retrieving the details of the currently logged in user. Added `GetCurrentUserQuery` and `GetCurrentUserQueryHandler` for the handling of the query. The DatabaseContext has been updated to include 'IManagementDbContext'. Other minor updates include fixing typos in endpoint descriptions, changing the comparison approach for domain events count, and updating several dependencies.