Removed GetInstitution test endpoint + created new migrations
The GetInstitutionEndpoint has been deleted as it's no longer needed. Also, changes were made to the Institutions entity in the database: the migration file has been created and snapshot updated. References to the newly renamed Institutions table have been updated in the database extension file and Institution repository has been added to it.