Remove ResultStatus class and refactor code for brevity
The ResultStatus class has been deleted and corresponding areas in the code have been refactored. This includes the adjustment of messages in the Result class and eliminating unnecessary conditions in DatabaseInitializer.cs. Further, the 'Status' property in the Result class was replaced with a 'State' property and its occurrences in the code have been updated.