Added institutionIds list to User
A new list of InstitutionId, '_institutionIds', has been introduced in the User class. This change affected the constructor of the User as well as the derived classes InstitutionUser and BackOfficeUser, where InstitutionId list is now part of the constructor. This allows keeping track of multiple institutions associated with a user.