Added Outbox functionality
This commit introduces a new Outbox service, complete with database migrations for the corresponding OutboxMessages. This service is designed to handle outgoing messages in an efficient and reliable manner. As part of this commit, configuration changes have been made to several files to properly integrate the new service, and additional methods were added to the OutboxWriter and MessageBus classes to improve functionality. A new file, OutboxMessageProcessorJob, has also been created for asynchronous message processing. Minor TODO indications were added and addressed in UserConfiguration and OutboxDbContext respectively. Finally, a new implementation for a switch-like functionality was introduced into the Common Utilities.