Added join extension method and updated access token claims
A new extension method 'Join' has been added to the EnumerableExtensions.cs to join the elements of a sequence into a string using a separator. This method is now utilized in the AccessTokenProvider.cs file to combine user's institution IDs instead of adding a TODO placeholder. The outdated code for disordering a list in EnumerableExtensions.cs file has been also simplified.