Added CourseConfiguration and CourseIdConverter + CourseCodeConverter
Added new 'CourseIdConverter' and 'CourseCodeConverter' to handle data conversions for the education module. Also, made adjustments to the 'Course.cs' file to change 'StartDate' and 'EndDate' to 'PlannedStart' and 'PlannedEnd'. Updated the database configurations in the 'CourseConfiguration.cs', 'CourseIdConverter.cs', and 'CourseCodeConverter.cs' files to use the new converters and reflect the changes made to the 'Course.cs' properties.