Added CreateCourse feature
The commit includes several changes that aim to add the functionality of creating a new course. It adds a new function ExistAsync in the IStudentRepository interface, creates an ICourseRepository interface, and implements it in a newly created CourseRepository class. Also, the Course class has been modified with additional properties, and handling methods for creating a new course have been created in the CreateCourseCommand and CreateCourseCommandHandler classes.