Added Scalar.AspNetCore dependency and update global sdk version
The code has been updated to include the Scalar.AspNetCore namespace in the Enscool.Bootstrapper program. Additionally, a new file, global.json, has been added to specify the .NET SDK version as 8.0.5 and enable the use of prerelease versions. The OpenApi library version in the Core.Infrastructure project has been upgraded to 8.0.5, and Scalar.AspNetCore has been added as a new package reference.