Removed outdated files and added UpdateInstitutionUser feature
Three handler classes, 'BrowseInstitutionUsersQueryHandler', 'CreateInstitutionUserCommandHandler', and 'RegisterCommandHandler', were removed as part of an earlier code refactoring. In addition, a new handler and corresponding command and request classes were created to handle the update functionality for an 'InstitutionUser'. Minor changes were also made to facilitate these updates, including modifying a user's 'Phone' attribute and 'Fullname' to be protected instead of private.